+++ src/NetStandard/E.Standard.WebGIS.Tools/Export/MapSeriesPrint.cs
@@ -2,6 +2,7 @@
 using E.Standard.Localization.Abstractions;
 using E.Standard.WebGIS.Core.Reflection;
 using E.Standard.WebGIS.Tools.Export.Calc;
+using E.Standard.WebGIS.Tools.Export.Exeption;
 using E.Standard.WebGIS.Tools.Export.Extensions;
 using E.Standard.WebGIS.Tools.Export.Models;
 using E.Standard.WebGIS.Tools.Extensions;
@@ -689,13 +690,24 @@ internal class MapSeriesPrint : IApiServerToolLocalizable<MapSeriesPrint>,
 
         var seriesCreator = new SeriesCreator(features, pageWidth, pageHeight, overlappingPercent);
 
-        var sketch = seriesType switch
+        MultiPoint sketch = new();
+        try
         {
-            SeriesType.AlongPolylines => seriesCreator.SeriesAlongPolylines(),
-            SeriesType.IntersectionRaster => seriesCreator.IntersectionRaster(),
-            SeriesType.BoundingBoxRaster => seriesCreator.BoundingBoxRaster(),
-            _ => throw new NotSupportedException($"Series type '{seriesType}' is not supported.")
-        };
+            int maxIterations = e.GetMaxMapSeriesPages() * 10;
+            sketch = seriesType switch
+            {
+                SeriesType.AlongPolylines => seriesCreator.SeriesAlongPolylines(maxIterations),
+                SeriesType.IntersectionRaster => seriesCreator.IntersectionRaster(maxIterations),
+                SeriesType.BoundingBoxRaster => seriesCreator.BoundingBoxRaster(maxIterations),
+                _ => throw new NotSupportedException($"Series type '{seriesType}' is not supported.")
+            };
+        }
+        catch (MapSeriesPrintToManyPagesExeption ex)
+        {
+            e[CreateSeriesValidationErrors] = String.Format(localizer.Localize(
+                "create-series-from-features.exception-to-many-iterations:body"),
+                                    ex.Iterations);
+        }
 
         // Set parameters for map series print for recalc rects etc...
         e[MapSeriesPrintLayoutId] = layoutId;